No internal worker class
